Paging bug fixed
- Basically the load_page_directory function was causing a page fault.
- The paging initialisation function now uses the bitmap allocator.
- a new heap allocator should be implemented
- and the physical address of the page directory is now correct. Because the paging_init function uses a translator function with the bitmap

+++ include/paging.h
@@ -1,7 +1,8 @@
 #define PDE_NUM 3 // page directory entries
 #define PTE_NUM 1024
+#include "stdint.h"
 
-extern unsigned int *page_directory;
+extern uintptr_t  *page_directory;
 
 void paging_init();
 int create_page_entry(int, char, char, char, char, char, char, char, char);
